services/api/requirements.txt
Django>=4.2,<6
requests>=2.32.0
django-cors-headers>=4.4.0
djangorestframework
feedparser
google-genai>=0.6.0
djangorestframework-simplejwt==5.3.1